backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python
supervision
numpy
deepface
tf-keras
django>=5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
pillow>=10.0.0